AT: improve Show and NFData instances
If the tensor in is AT representation (Accelerate AST) this will evaluate the program to normal form. Unfortunately the input term will still be in AT form, so the work of evaluating the AST to real values is sort of wasted, but at least the instances now do what they advertise on the tin.